Move data model types into dedicated header
